This change avoids a warning on the Visual C++ compiler.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/simgrid/simgrid/trunk@4984 48e7efb5-ca39-0410-a469-dd3cf9ba447f

+++ include/surf/surfxml_parse.h
@@ -61,7 +61,7 @@ XBT_PUBLIC_DATA(int_f_void_t) surf_parse; /* Entry-point to the parser. Set this
 #define SURFXML_BUFFER_SET(key,val) do { \
   AX_surfxml_##key=AX_ptr; \
   strcpy(A_surfxml_##key,val); \
-  AX_ptr+=strlen(val)+1; } while(0)
+  AX_ptr+=(int)strlen(val)+1; } while(0)
 
 #define SURFXML_BUFFER_RESET() do { \
   AX_ptr = 0; \
